+++ aws/services/gen_mocks_test.go
@@ -19,6 +19,7 @@ limitations under the License.
 package awsservices
 
 import (
+	"context"
 	"strconv"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -104,7 +105,7 @@ func (m *mockEc2) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockEc2) FetchResources() (*graph.Graph, error) {
+func (m *mockEc2) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -112,7 +113,7 @@ func (m *mockEc2) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockEc2) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockEc2) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -232,7 +233,7 @@ func (m *mockElbv2) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockElbv2) FetchResources() (*graph.Graph, error) {
+func (m *mockElbv2) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -240,7 +241,7 @@ func (m *mockElbv2) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockElbv2) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockElbv2) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -297,7 +298,7 @@ func (m *mockRds) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockRds) FetchResources() (*graph.Graph, error) {
+func (m *mockRds) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -305,7 +306,7 @@ func (m *mockRds) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockRds) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockRds) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -376,7 +377,7 @@ func (m *mockAutoscaling) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockAutoscaling) FetchResources() (*graph.Graph, error) {
+func (m *mockAutoscaling) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -384,7 +385,7 @@ func (m *mockAutoscaling) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockAutoscaling) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockAutoscaling) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -477,7 +478,7 @@ func (m *mockIam) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockIam) FetchResources() (*graph.Graph, error) {
+func (m *mockIam) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -485,7 +486,7 @@ func (m *mockIam) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockIam) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockIam) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -556,7 +557,7 @@ func (m *mockS3) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockS3) FetchResources() (*graph.Graph, error) {
+func (m *mockS3) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -564,7 +565,7 @@ func (m *mockS3) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockS3) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockS3) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -600,7 +601,7 @@ func (m *mockSns) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockSns) FetchResources() (*graph.Graph, error) {
+func (m *mockSns) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -608,7 +609,7 @@ func (m *mockSns) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockSns) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockSns) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -678,7 +679,7 @@ func (m *mockSqs) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockSqs) FetchResources() (*graph.Graph, error) {
+func (m *mockSqs) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -686,7 +687,7 @@ func (m *mockSqs) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockSqs) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockSqs) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -726,7 +727,7 @@ func (m *mockRoute53) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockRoute53) FetchResources() (*graph.Graph, error) {
+func (m *mockRoute53) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -734,7 +735,7 @@ func (m *mockRoute53) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockRoute53) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockRoute53) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -786,7 +787,7 @@ func (m *mockLambda) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockLambda) FetchResources() (*graph.Graph, error) {
+func (m *mockLambda) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -794,7 +795,7 @@ func (m *mockLambda) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockLambda) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockLambda) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -847,7 +848,7 @@ func (m *mockCloudwatch) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockCloudwatch) FetchResources() (*graph.Graph, error) {
+func (m *mockCloudwatch) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -855,7 +856,7 @@ func (m *mockCloudwatch) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockCloudwatch) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockCloudwatch) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -924,7 +925,7 @@ func (m *mockCloudfront) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockCloudfront) FetchResources() (*graph.Graph, error) {
+func (m *mockCloudfront) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -932,7 +933,7 @@ func (m *mockCloudfront) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockCloudfront) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockCloudfront) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -967,7 +968,7 @@ func (m *mockCloudformation) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockCloudformation) FetchResources() (*graph.Graph, error) {
+func (m *mockCloudformation) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -975,7 +976,7 @@ func (m *mockCloudformation) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockCloudformation) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockCloudformation) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -1027,7 +1028,7 @@ func (m *mockEcr) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockEcr) FetchResources() (*graph.Graph, error) {
+func (m *mockEcr) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -1035,7 +1036,7 @@ func (m *mockEcr) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockEcr) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockEcr) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -1094,7 +1095,7 @@ func (m *mockEcs) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m *mockEcs) FetchResources() (*graph.Graph, error) {
+func (m *mockEcs) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -1102,7 +1103,7 @@ func (m *mockEcs) IsSyncDisabled() bool {
 	return false
 }
 
-func (m *mockEcs) FetchByType(t string) (*graph.Graph, error) {
+func (m *mockEcs) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }