+++ gen/aws/generators/mocks.go
@@ -117,7 +117,7 @@ func (m * {{ $mock.Name }}) ResourceTypes() []string {
 	return []string{}
 }
 
-func (m * {{ $mock.Name }}) FetchResources() (*graph.Graph, error) {
+func (m * {{ $mock.Name }}) Fetch(context.Context) (*graph.Graph, error) {
 	return nil, nil
 }
 
@@ -125,7 +125,7 @@ func (m * {{ $mock.Name }}) IsSyncDisabled() bool {
 	return false
 }
 
-func (m * {{ $mock.Name }}) FetchByType(t string) (*graph.Graph, error) {
+func (m * {{ $mock.Name }}) FetchByType(context.Context, string) (*graph.Graph, error) {
 	return nil, nil
 }